requirements.txt
fastapi[all]==0.63.0
email-validator==1.1.2
uvicorn==0.13.2
PyJWT==2.0.0
pymongo==3.11.2
python-multipart==0.0.5
passlib[bcrypt]==1.7.4
motor==2.3.0
databases==0.4.1
mailjet-rest==1.3.4
gunicorn==20.0.4
boto3==1.16.43
cryptography==3.3.1
twilio==6.50.1
numpy==1.19.4
opencv-python==4.4.0.46
scipy==1.5.4
aiofiles==0.6.0
stripe==2.55.1
firebase-admin==4.5.0
paho-mqtt==1.5.1
pytest==6.2.1
